Clarify maximum message size behavior
This is confusing only because you can allow NewMessenger to choose for
you, essentially. So for now, the guarantee is that if you pick 0,
Messenger's upper limit is reasonably large, and maybe unreasonably
large, but it could change depending on the target system's limits. It
doesn't right now, but I'd like to eventually add a check for e.g.,
whether it's even possible to encode your max message size on the
architecture you're building for.